ref: rules and faq
The rules of NaN are split between two channels: #regles and #faq. The
\#regle channel just list 8 rules that are commonly found on discord
servers : be nice to other, spell correctly, don't send ads, etc. The
\#faq channel lists various situation the newbies can find themselves in
and explain how to correctly behave in those situations.

The problem is that the #regle channel doesn't only states rules: it
also explain how things like the :modo: flag works. Same goes for the
\#faq channel, some of the "how to behave" sitation are rules de facto.

We argue that the two channels should be merged into one into a rule book
that both list rules and explain how the server works.

We identified 4 categories:

1/ The rules, classic set of rules common with a lot a discord servers.
2/ How the server works, how to join channels, what channels are for,
   how to ask for help, etc.
3/ The moderated channels: #offre, #vos-projets, #lives.
4/ Some meta informations about the server like how to suggest changes.

By merging all the rules in one place, we had the opportunity to
carefully reword every sentence so it better reflects the server as it
is in 2021.

Co-authored-by: Lezenn <51204940+GiregL@users.noreply.github.com>
Closes #13