add "Culte du Code" in the security guideline
"Code du Code" is a french community where it is possible to discuss about
security. They offer an environment that is proper ethical-wise and that fit
NaN's opinion on what should a security communty looks like and behave.

Until we have our own workforce capable of enforcing such ethical
considerations in a dedicated channel. Advertising another discord server
is a good middle ground.