add: !courspython fluent book
Add the *lizard book* in the list of books we recommend for Python.

This is a huge, 980 pages long, book "written for practicing Python programmers who want to become proficient in Python 3". This is the most rented book on Python from our library at work. I had myself the opportunity to rent it one week and just reading the first few chapters convinced me to buy it for my own.

I think it deserves to be ranked first in the list of books we recommend, above the Cookbook (that I still consider to be excellent).